+++ src/config/parseSelector.ts
@@ -1,28 +1,27 @@
-import { Selector, Config, RawConfig } from './types';
+import { Selector, RawConfig } from './types';
 
-const parseSelector = (selector: Selector): RawConfig => {
-  const config: Config = {};
+function parseSelector<Initial = unknown>(
+  selector: Selector
+): RawConfig<Initial> {
+  const config: RawConfig<Initial> = { selector: '' };
 
-  let $selector, methods, attr;
-
-  if (Array.isArray(selector)) {
-    return { selector: selector.join(', ') };
-  }
+  let $selector: string, methods: string[], attr: string;
 
   if (typeof selector !== 'string') {
     return selector;
   }
 
-  if (selector?.includes('|')) {
+  if (selector.includes('|')) {
     [$selector, ...methods] = selector.split('|').map((key) => key.trim());
-    methods?.map((p) => p.trim());
+
+    methods = methods.map((p: string) => p.trim());
   }
 
   if (!$selector) {
     $selector = selector;
   }
 
-  if ($selector?.includes('@')) {
+  if ($selector.includes('@')) {
     [$selector, attr] = $selector.split('@').map((key) => key.trim());
   }
 
@@ -41,6 +40,6 @@ const parseSelector = (selector: Selector): RawConfig => {
   }
 
   return config;
-};
+}
 
 export default parseSelector;