+++ src/parser/parse.ts
@@ -1,10 +1,12 @@
 import { load, CheerioAPI } from 'cheerio';
-import { InputConfig } from '../config/types';
+import parseSelector from '../config/parseSelector';
+import { Config } from '../config/types';
+import { getRawConfig } from './getRawConfig';
 import getValue from './getValue';
 
-function parse(
+function parse<Initial = unknown>(
   data: string | CheerioAPI,
-  config: InputConfig
+  config: Config<Initial>
 ): Record<string, unknown> {
   let $;
 
@@ -14,7 +16,21 @@ function parse(
     $ = data;
   }
 
-  return getValue({ $ }, config);
+  const rawConf = getRawConfig(config);
+
+  if (Array.isArray(rawConf)) {
+    for (const conf of rawConf) {
+      const value = getValue({ $, el: $ }, conf);
+
+      if (value !== null) {
+        return value;
+      }
+    }
+
+    return null;
+  }
+
+  return getValue({ $ }, rawConf);
 }
 
 export default parse;