+++ src/parser/transformValue.ts
@@ -1,16 +1,20 @@
-import { Config } from '../config/types';
+import { RawConfig } from '../config/types';
 import Methods from './methods';
 import execRegex from './regex/execRegex';
+import { Value } from './value';
 
-function transformValue(value: any, config: Config) {
+function transformValue<Initial = unknown>(
+  value: Value<Initial>,
+  config: RawConfig<Initial>
+) {
   const { trim, regex, type, methods = [], transform } = config;
 
   if (typeof value === 'string' && trim !== false) {
     value = value.trim();
-  }
 
-  if (regex) {
-    value = execRegex(value, regex);
+    if (regex) {
+      value = execRegex(value, regex);
+    }
   }
 
   if (type) {