+++ src/parser/value.ts
@@ -1,0 +1,1 @@
+export type Value<Initial = unknown> = string | Initial;